implementando viewmodel
