+++ app/src/main/java/com/br/matchmovies/viewmodel/MovieDetailsViewModel.kt
@@ -1,0 +1,6 @@
+package com.br.matchmovies.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class MovieDetailsViewModel : ViewModel() {
+}
